+++ CHANGELOG.md
@@ -19,6 +19,9 @@ All Notable changes to `PHP Telegram Bot Api` will be documented in this file
 - Add `\TelegramBot\Api\Types\Update::$messageReaction` field
 - Add `\TelegramBot\Api\Types\Update::$messageReactionCount` field
 - Add `\TelegramBot\Api\BotApi::setMessageReaction` api method
+- Add `\TelegramBot\Api\BotApi::deleteMessages` api method
+- Add `\TelegramBot\Api\BotApi::copyMessages` api method
+- Add `\TelegramBot\Api\BotApi::forwardMessages` api method
 
 ### Deprecated
 - Deprecate `reply_to_message_id` and `allow_sending_without_reply` parameters to `\TelegramBot\Api\BotApi` methods. Use `reply_parameters` instead.